# Build ESM and CJS bundles simultaneously