--- a/.changeset/breezy-mirrors-behave.md
+++ b/.changeset/breezy-mirrors-behave.md
@@ -1,0 +1,5 @@
+---
+'@crackle/core': patch
+---
+
+Upgrade Vite and Rollup to latest