--- a/.changeset/popular-radios-bathe.md
+++ b/.changeset/popular-radios-bathe.md
@@ -1,0 +1,5 @@
+---
+'@crackle/core': minor
+---
+
+Run DTS build once for all entries