--- a/.changeset/violet-spiders-cover.md
+++ b/.changeset/violet-spiders-cover.md
@@ -1,0 +1,5 @@
+---
+'@crackle/core': minor
+---
+
+Build ESM and CJS bundles simultaneously