--- a/packages/core/package.json
+++ b/packages/core/package.json
@@ -113,7 +113,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "resolve-from": "^5.0.0",
-    "rollup": "^3.10.0",
+    "rollup": "^3.18.0",
     "rollup-plugin-dts": "^5.1.1",
     "rollup-plugin-node-externals": "^5.0.2",
     "semver": "^7.3.8",
@@ -122,7 +122,7 @@
     "sort-package-json": "^1.57.0",
     "type-fest": "^3.2.0",
     "used-styles": "^2.4.1",
-    "vite": "^4.1.1"
+    "vite": "^4.1.4"
   },
   "devDependencies": {
     "@types/babel__core": "^7.1.20",