--- a/packages/core/src/types.ts
+++ b/packages/core/src/types.ts
@@ -1,19 +1,8 @@
 import type React from 'react';
-import type { InlineConfig } from 'vite';
 
 export type { PackageJson } from 'type-fest';
 
 export type Format = 'esm' | 'cjs' | 'dts';
-
-export type GetArrayType<T> = T extends Array<infer U> ? U : never;
-
-export type ManualChunksFn = NonNullable<
-  GetArrayType<
-    NonNullable<
-      NonNullable<NonNullable<InlineConfig['build']>['rollupOptions']>['output']
-    >
-  >['manualChunks']
->;
 
 type RoutePath = string;
 export type AppShell<MetadataType extends Record<string, any> | void = void> =