--- a/packages/core/src/vite-config.ts
+++ b/packages/core/src/vite-config.ts
@@ -1,6 +1,12 @@
+import assert from 'assert';
+
+import type { OutputOptions } from 'rollup';
 import type { InlineConfig } from 'vite';
 
 import type { EnhancedConfig } from './config';
+import { distDir } from './constants';
+import type { Format, PackageEntryPoint } from './types';
+import { extensionForFormat } from './utils/files';
 
 export const commonViteConfig = (config: EnhancedConfig): InlineConfig => ({
   root: config.root,
@@ -14,3 +20,27 @@ export const commonViteConfig = (config: EnhancedConfig): InlineConfig => ({
     global: JSON.stringify({}),
   },
 });
+
+export const commonOutputOptions = (
+  config: EnhancedConfig,
+  entryPoints: PackageEntryPoint[],
+  format: Format = 'esm',
+) => {
+  const extension = extensionForFormat(format);
+
+  return {
+    dir: distDir,
+    hoistTransitiveImports: false,
+    minifyInternalExports: false,
+    entryFileNames(chunkInfo) {
+      const entry = entryPoints.find(
+        ({ entryPath }) => chunkInfo.facadeModuleId === entryPath,
+      );
+      assert(entry, `entry not found for ${chunkInfo.facadeModuleId}`);
+
+      return entry.getOutputPath(format);
+    },
+    chunkFileNames: ({ name }) =>
+      name.endsWith(extension) ? name : `${name}.chunk.${extension}`,
+  } satisfies OutputOptions;
+};