- Upgrade Vite and Rollup to latest
- Build ESM and CJS bundles simultaneously
- Run DTS build once for all entries. Previously each entry was built separately to avoid having `.d.ts` chunks, but now we can use [`experimentalMinChunkSize`](https://rollupjs.org/configuration-options/#output-experimentalminchunksize) to achieve the same result

### Benchmark results (packaging Braid)

`master`:
```
➜ hyperfine "nr build"
Benchmark 1: nr build
  Time (mean ± σ):     28.016 s ±  0.630 s    [User: 39.951 s, System: 5.235 s]
  Range (min … max):   26.568 s … 28.737 s    10 runs
```

This branch:
```
➜ hyperfine "nr build"                                     
Benchmark 1: nr build
  Time (mean ± σ):     13.141 s ±  0.448 s    [User: 19.233 s, System: 2.623 s]
  Range (min … max):   12.614 s … 14.199 s    10 runs
```